requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
anthropic==0.40.0
langchain==0.1.0
langgraph==0.2.45
chromadb==0.5.23
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.1
